Extract SerializerHandlerBenchmarksBase
